Reproduce error with numeric return types
When a python snippet returns a numeric type, we get a crash.

This is a snippet plugin.  It probably makes sense to stringify any
output before manipulating it.  It probably also makes sense to add
tests to cover this newly discovered edge case, because, well, python.

```
An error occured. This is either a bug in UltiSnips or a bug in a
snippet definition. If you think this is a bug, please report it to
https://github.com/SirVer/ultisnips/issues/new
Please read and follow:
https://github.com/SirVer/ultisnips/blob/master/CONTRIBUTING.md#reproducing-bugs

Following is the full stack trace:
Traceback (most recent call last):
  File "/src/UltiSnips/pythonx/UltiSnips/err_to_scratch_buffer.py", line 18, in wrapper
    return func(self, *args, **kwds)
  File "/src/UltiSnips/pythonx/UltiSnips/snippet_manager.py", line 127, in expand
    if not self._try_expand():
  File "/src/UltiSnips/pythonx/UltiSnips/snippet_manager.py", line 759, in _try_expand
    self._do_snippet(snippet, before)
  File "/src/UltiSnips/pythonx/UltiSnips/snippet_manager.py", line 712, in _do_snippet
    text_before, self._visual_content, parent, start, end
  File "/src/UltiSnips/pythonx/UltiSnips/snippet/definition/base.py", line 488, in launch
    snippet_instance.update_textobjects(vim_helper.buf)
  File "/src/UltiSnips/pythonx/UltiSnips/text_objects/snippet_instance.py", line 93, in update_textobjects
    if obj._update(done, buf):
  File "/src/UltiSnips/pythonx/UltiSnips/text_objects/python_code.py", line 279, in _update
    self.overwrite(buf, rv)
  File "/src/UltiSnips/pythonx/UltiSnips/text_objects/base.py", line 146, in overwrite
    self._end = _replace_text(buf, self._start, self._end, gtext)
  File "/src/UltiSnips/pythonx/UltiSnips/text_objects/base.py", line 22, in _replace_text
    lines = text.split("\n")
AttributeError: 'int' object has no attribute 'split'
```